chore(zql): address review comments
from #1854

- https://github.com/rocicorp/mono/pull/1854#discussion_r1606937855
- https://github.com/rocicorp/mono/pull/1854#discussion_r1606916649
- https://github.com/rocicorp/mono/pull/1854#discussion_r1606896834
- https://github.com/rocicorp/mono/pull/1854#discussion_r1606897814
- https://github.com/rocicorp/mono/pull/1854#discussion_r1606902285
- https://github.com/rocicorp/mono/pull/1854#discussion_r1606903520
